FieldAPI: Rename Field API modules in `parallel` and `data_offload`
