+++ loki/transformations/data_offload/__init__.py
@@ -8,6 +8,6 @@
 Sub-package providing data offload transformations.
 """
 
-from loki.transformations.data_offload.field_api import * # noqa
+from loki.transformations.data_offload.field_offload import * # noqa
 from loki.transformations.data_offload.global_var import * # noqa
 from loki.transformations.data_offload.offload import * # noqa
